chore(docs): move fluentd config to separate file
