+++ docs/configuration/crds/v1beta1/fluentd_types.md
@@ -172,60 +172,6 @@ Available in Logging operator version 4.5 and later. Configure sidecar container
 
 
 
----
-title: FluentdConfig
-weight: 200
-generated_file: true
----
-
-## FluentdConfig
-
-FluentdConfig
-
-###  (metav1.TypeMeta, required) {#fluentdconfig-}
-
-
-### metadata (metav1.ObjectMeta, optional) {#fluentdconfig-metadata}
-
-
-### spec (FluentdSpec, optional) {#fluentdconfig-spec}
-
-
-### status (FluentdConfigStatus, optional) {#fluentdconfig-status}
-
-
-
-## FluentdConfigStatus
-
-FluentdConfigStatus
-
-### active (*bool, optional) {#fluentdconfigstatus-active}
-
-
-### logging (string, optional) {#fluentdconfigstatus-logging}
-
-
-### problems ([]string, optional) {#fluentdconfigstatus-problems}
-
-
-### problemsCount (int, optional) {#fluentdconfigstatus-problemscount}
-
-
-
-## FluentdConfigList
-
-FluentdConfigList
-
-###  (metav1.TypeMeta, required) {#fluentdconfiglist-}
-
-
-### metadata (metav1.ListMeta, optional) {#fluentdconfiglist-metadata}
-
-
-### items ([]FluentdConfig, required) {#fluentdconfiglist-items}
-
-
-
 ## FluentOutLogrotate
 
 ### age (string, optional) {#fluentoutlogrotate-age}
